WIP use annotated entities - audit entities
Hibernate 5 allows for (and encourages) annotations instead of XML
configuration. Start doing that for audit records and audit details,
which are one of the first things that were failing. Use generated IDs
instead of manually-assigned IDs.

I am not yet setting any character limits on the columns, or enforcing
names - AFAIK these tables are not referenced anywhere else, so the
automatically inferred names are fine.