use UUIDs for new users' IDs
Previously, the app was generating user IDs like
`XXXXCMSO00000000000000000001`. It's not clear to me why they were doing
that, or even why it is useful to have both a user ID and a user name.

For now, replace those app-generated alpha-numeric user IDs with
Hibernate-generated UUIDs (stored as a string). Maybe later we can
revisit this decision. This allows self-registration!